Create Two Functions in Project.js
minimal syntax code on dom.js, the two Functions
serve to save the projects and their respective
todos,title, and desc. It uses the local storage
function its kind of like getter and setter.
I set the objects and stringify each project
in the project arr and get them through parse
afterwards to save its implementation is abstracted
to be ntoiced if theres problem on implementation